requirements.txt
selenium==4.18.1
fpdf2==2.7.8
requests==2.31.0
python-slugify==8.0.4
prettyconf==2.2.1
logzero==1.7.0
typer==0.11.0
